added subtypes & tests. resolved comments
